refactor getByPath to internal
