+++ get.js
@@ -1,78 +1,4 @@
-const isArray = require('./_internal/isArray')
-const memoizeCappedUnary = require('./_internal/memoizeCappedUnary')
-
-// a[0].b.c
-const pathStringSplitRegex = /[.|[|\]]+/
-
-/**
- * @name pathStringSplit
- *
- * @synopsis
- * ```coffeescript [specscript]
- * pathStringSplit(pathString string) -> Array<string>
- * ```
- *
- * @note
- * a[0].b.c
- * a.0.b[0][1].c
- */
-const pathStringSplit = function (pathString) {
-  const pathStringLastIndex = pathString.length - 1,
-    firstChar = pathString[0],
-    lastChar = pathString[pathStringLastIndex],
-    isFirstCharLeftBracket = firstChar == '[',
-    isLastCharRightBracket = lastChar == ']'
-
-  if (isFirstCharLeftBracket && isLastCharRightBracket) {
-    return pathString.slice(1, pathStringLastIndex).split(pathStringSplitRegex)
-  } else if (isFirstCharLeftBracket) {
-    return pathString.slice(1).split(pathStringSplitRegex)
-  } else if (isLastCharRightBracket) {
-    return pathString.slice(0, pathStringLastIndex).split(pathStringSplitRegex)
-  }
-  return pathString.split(pathStringSplitRegex)
-}
-
-// memoized version of pathStringSplit, max cache size 500
-const memoizedCappedPathStringSplit = memoizeCappedUnary(pathStringSplit, 500)
-
-/**
- * @name pathToArray
- *
- * @synopsis
- * ```coffeescript [specscript]
- * pathToArray(path string|number|Array<string|number>) -> Array
- * ```
- */
-const pathToArray = path => isArray(path) ? path
-  : typeof path == 'string' ? memoizedCappedPathStringSplit(path)
-  : [path]
-
-/**
- * @name getByPath
- *
- * @synopsis
- * ```coffeescript [specscript]
- * getByPath<
- *   value any,
- *   path string|number|Array<string|number>,
- * >(value, path) -> valueAtPath any
- * ```
- */
-const getByPath = function (value, path) {
-  const pathArray = pathToArray(path),
-    length = pathArray.length
-  let index = -1,
-    result = value
-  while (++index < length) {
-    result = result[pathArray[index]]
-    if (result == null) {
-      return undefined
-    }
-  }
-  return result
-}
-
+const getByPath = require('./_internal/getByPath')
 
 /**
  * @name get
